Arregle problema al iniciar
