+++ app/src/main/java/placesafe/placesafe/OpinionController.java
@@ -1,7 +1,6 @@
 package placesafe.placesafe;
 
 import android.app.Application;
-import android.content.Context;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
@@ -39,40 +38,5 @@ public class OpinionController {
         }, data);
 
         opinion.setText("");
-    }
-    public static void  fillOpinions(Application app,final View v,final Context context){
-
-        RequestVolley req = RequestVolley.getInstance(app.getApplicationContext());
-
-        req.requestString("POST", "/opinions", new Response.Listener<String>() {
-            @Override
-            public void onResponse(String responser) {
-
-                try {
-                    List<Opinion> opinions = new ArrayList<>();
-                    RecyclerView rv = (RecyclerView) v.findViewById(R.id.opinionesList);
-                    LinearLayoutManager lym = new LinearLayoutManager(context);
-                    rv.setLayoutManager(lym);
-
-
-                    JSONArray response = new JSONArray(responser);
-                    for (int i = 0; i < response.length(); i++) {
-                        JSONObject jresponse = response.getJSONObject(i);
-                        String nickname = jresponse.getString("nickname");
-                        String opinionText = jresponse.getString("opinionText");
-                        opinions.add(new Opinion(nickname, opinionText));
-                    }
-
-                    AdapterOpinions adapterOpinions = new AdapterOpinions(opinions);
-                    rv.setAdapter(adapterOpinions);
-
-
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-
     }
 }
